Move SafeEnumeratorHandle
