+++ source/icu.net/Collation/RuleBasedCollator.cs
@@ -313,54 +313,6 @@ namespace Icu.Collation
 			return locales;
 		}
 
-		internal sealed class SafeEnumeratorHandle : SafeHandle
-		{
-			public SafeEnumeratorHandle()
-				:
-					base(IntPtr.Zero, true) { }
-
-			///<summary>
-			///When overridden in a derived class, executes the code required to free the handle.
-			///</summary>
-			///<returns>
-			///true if the handle is released successfully; otherwise, in the event of a catastrophic failure, false. In this case, it generates a ReleaseHandleFailed Managed Debugging Assistant.
-			///</returns>
-#if !NETSTANDARD1_6
-			[ReliabilityContract(Consistency.WillNotCorruptState, Cer.Success)]
-#endif
-			protected override bool ReleaseHandle()
-			{
-				NativeMethods.uenum_close(handle);
-				handle = IntPtr.Zero;
-				return true;
-			}
-
-			///<summary>
-			///When overridden in a derived class, gets a value indicating whether the handle value is invalid.
-			///</summary>
-			///<returns>
-			///true if the handle is valid; otherwise, false.
-			///</returns>
-			public override bool IsInvalid
-			{
-				get { return (handle == IntPtr.Zero); }
-			}
-
-			public string Next()
-			{
-				ErrorCode e;
-				int length;
-				IntPtr str = NativeMethods.uenum_unext(this, out length, out e);
-				if (str == IntPtr.Zero)
-				{
-					return null;
-				}
-				string result = Marshal.PtrToStringUni(str, length);
-				ExceptionFromErrorCode.ThrowIfError(e);
-				return result;
-			}
-		}
-
 		#region ICloneable Members
 
 		///<summary>
